O tipo do valor dos produtos foi trocado para Float no banco de dados e no model.

+++ src/main/java/com/squares/products/dtos/ProductRecordDTO.java
@@ -3,7 +3,5 @@ package com.squares.products.dtos;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
-import java.math.BigDecimal;
-
-public record ProductRecordDTO(@NotBlank String name, @NotNull BigDecimal value) {
+public record ProductRecordDTO(@NotBlank String name, @NotNull Float value) {
 }
